add seed file

+++ db/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE department (
 CREATE TABLE roles (
     id INTEGER AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR (30),
-    salary DECIMAL (50,5),
+    salary DECIMAL (65,5),
     department_id INTEGER,
     CONSTRAINT department_id FOREIGN KEY(department_id) REFERENCES department(id) ON DELETE CASCADE
 );
@@ -21,5 +21,5 @@ CREATE TABLE employee (
     last_name VARCHAR (30) NOT NULL,
     role_id INTEGER,
     CONSTRAINT role_id FOREIGN KEY(role_id) REFERENCES roles(id) ON DELETE CASCADE,
-    manager_id INTEGER(3)
+    manager_id INTEGER(4)
 );
